+++ source/B2B.Transactions/OutgoingMessages/AcceptMessage.cs
@@ -11,16 +11,11 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-using System;
-
-namespace B2B.Transactions.Tests
+namespace B2B.Transactions.OutgoingMessages
 {
-    public static class MessageIdGenerator
+    public class AcceptMessage
     {
-        public static string Generate()
-        {
-            return Guid.NewGuid().ToString();
-        }
+        #pragma warning disable
+        public string MessagePayload { get; init; }
     }
 }
